bib: suppress warnings from logrus
Suppresses the message that we _always_ get:

  WARN[0000] Failed to load consumer certs: no consumer key found

This is a warning from images that indicates that the host isn't
subscribed or the subscription certs couldn't be read.  It's useful for
the specific case of running osbuild-composer on RHEL and is meant to be
logged in the journal.  Here, it's just noise.