cmd/bootc-image-builder: upload AMI when flags are specified
When building an AMI with bootc-image-builder, if the AWS flags are
defined (they are all required together, so one being set implies they
are all set), then upload and register the AMI image.

If the flags are defined for a non-AMI image type, error out.

Also, check and fail early if the credentials aren't set in the
environment, so we don't error out after the image is built.