vm: make AWS subclass of VM class
The AWS class manages a single EC2 instance from a given AMI.
Entering the context defined by the class (i.e. with AWS(id) ...)
performs the following actions (defined in start()):
- Create a security group that allows SSH logins from any IP.
- Create an instance from the given AMI ID and include the new security
  group.
- Block until the instance is running.

On exit, the instance is terminated and the security group deleted (in
that order).

Credentials aren't managed by the class.  They are assumed to be
managed externally (env vars).

Based heavily on examples found in the aws sdk docs:
https://github.com/awsdocs/aws-doc-sdk-examples/blob/afb3309b385db90e6ff01c0268429cde9610f53e/python/example_code/ec2/