Drain logic from entrypoint into Go code
This is prep for adding multiple entrypoints; right now the
Go code has a `build` verb that is always injected by this
wrapper.

Move the default values into the Go code.

A future change will likely make things so that users are
required to pass `build` (or perhaps `build-disk`) and
this shell script goes away entirely.