test: remove AMI test cases from "direct-boot"
The AMI tests are done via "ami-boot". Having parts in "direct-boot"
is confusing and also leads to bugs, i.e. we generate the ami in
the "direct-boot" test but then the ami is not uploaded to AWS but
just stored locally. This means that the boottest in AWS will fail
and also that the upload progress check will fail.

The ami-boot tests are done via tmt so there is no test coverage
regression here.