test: update tests to allow ssh keys
With ssh a root login is only possible via a sshkey. So let's
support this so that we can run `bootc status` which requires
root privs.
With the switch to bootc we need to adjust the testing. We inject
a root ssh key now and just use that for login.