bib: use target architecture for buildroot as well
The existing code is trying to fetch two containers when doing
a cross-arch image. The buildroot is run in the native architecture
and only the target tree is run non-native. However it turns out
this is not neccessary as even doing the buildroot in non-native
is fast enough (given that it's mostly I/O and all I/O is run
naively via syscall translation).

So this commit removes the fetching of the native arch container
for the buildroot. Note that in practise this was already the
case when passing --target-arch to any container that had native
build available. Here due to a bug/misfeature in the resolver library
used it already resolved (incorrectly) to the target architecture and
nobody noticed because it was good enough (after [0] got fixed).

[0] https://lists.nongnu.org/archive/html/qemu-devel/2024-02/msg03971.html