bib: always set an architecture in the go unit tests
Previously the `0` value of the images `arch.Architecture` iota
was of type `ARCH_AARCH64`. However with
https://github.com/osbuild/images/pull/585/commits/05355326188f3d2ad49003784d681524eba466ce
this is no longer the case an the zero type is now a proper
`UNSET` type. And the tests now fail because they did not set
the architecture.

This commit fixes this by setting the architecture.