README,bib: hide `--config` from users, remove from docs
With the default to pick up the config from `/config.{toml,json}`
nowdays the `--config` flag is not very useful anymore. It is
actually confusing users as they may hink they to pass two config
arguments into the container.

So this commit removes `--config` from the README and hides it
in bib. It is not removed entirely for two reasons:
1. backward compatibility with existing scripts
2. it is useful for developers to run
```
$ sudo ./bootc-image-builder manifest --config local-config.json
```
to get a manifest from outside a container for debugging purposes
(but it's not a strong reason).