bib: add (hidden) `bootc-image-builder version` command
This commit adds a (hidden) `version` command. It currently
only contains the git revision of build - that is useful for
easier debugging, i.e. we can now easily ask what revision
the user was using when a bug is reported.

Once we actually do real bib releases we could also include
the real version number.