test: drop size comparison in assert_fs_customizations
The size comparison in the build test is bad.  Customizations control
the size of a partition but the function was comparing the usable size
of a filesystem, which can be very different.  Attempts were made to
cover the difference, by simply reducing the expected size by a fixed
amount, but were ultimately inadequate.

We can do better.