cli: implement messy/untested "image-builder build"
This commit implements the `build` command. It will build an
image based on the given distro/type spec.

Note that no cross-arch building is possible (unlike `bib`) because
we would need to construct a cross-arch buildroot. This is requires
rpm/libc in the cross-arch first so there is a chicken-egg issue.
We *could* of course construct our buildroots from a container
base initially, then cross-building would work with qemu-user.