fixup! rework argument handling
This commit tweaks the argument handling based on the suggestion
by @achilleas-k (thanks!). Now it takes
```console
$ image-builder manifest centos-9 qcow2 ./path/to/blueprint
...
$ image-builder manifest --arch s390x centos-9 qcow2
...
```

This means it is no longer copy/paste friendly from `--list-images`
but we can tweak the output of that so that by default we only
how `distro:foo type:bar` instead of the full arch string and
hint via something like `isatty(0)` for interactive users that
there are more arches available.