bib: improve the "/" size calculation for LVM
This commit improves the rootfs size calculation for advanced
partitioning.  Just like for the tranditional filesystem
customization where we use `updateFilesystemSizes()` to ensure
that the rootfs is at least the two times size of the container
we need to do the same for the advanced disk partitioning.

The disk library will need a `RequiredMinSizes` parameter
(see gh#748) to know what sizes to pick. So far we did not
pass this which can leads to a "0B" root partition.

The naive fix was to just include a hardcoded:
```
requiredDirectorySizes = map[string]uint64{
        "/":    1 * datasizes.GiB,
        "/usr": 2 * datasizes.GiB,
}
```
but of course that is not ideal because we calculate the
size of disk based on 2x the container size. We set this
for the rootfs so the entire disk is big enough. However
because in advanced partitioning "/" is not automatically
expanded we end up with a potentially too small "/".

This commit tweaks the calculcation so that it sets the
requiredDirectorySizes for "/" to be at least 2x container
size.

Note that a custom "/usr" is not supported in image mode so splitting
rootfsMinSize between / and /usr is not a concern.