cmd: tweak/clarify the lifecycle of the progressbar
manifestFromCobra generate an osbuild manifest from a cobra
commandline and it also takes a progress.ProgressBar as input.

It needs an unstarted progres bar and will start it at the right
point (it cannot be started yet to avoid the "podman pull" progress
and our progress fighting). The caller is responsible for stopping
the progress bar (this functtion cannot know what else needs to happen
after manifest generation).

TODO: provide a podman progress reader to integrate the podman progress
into our progress.

Thanks to Ondrej for the suggestion to clarify the relationship
further.