cloud: temporary stop progress before uploading
The upload code is currently using a plain "pb.ProgressBar" and
this is tested as part of our integration test. This will collide
with our own progress implementation. For now workaround this by
stopping our own progress and let plain "pb" takeover.

I was tempted to fix it right away but the PR with the progress
is already relatively big so I opted for a followup.