spec: copy the embedded repo files from the images library
Now that we have merged images PR#1112 and have access to the
repository files via the vendor/ directory start with a very
simple spec update and copy the files into our own search
path. Note that once reporegistry is fixed we will not embed
the file in the rpm at all, we will just go:embed.

This also updates the search path for repositories to:
```
var defaultDataDirs = []string{
	"/etc/image-builder",
	"/usr/share/image-builder",
}
```