test: copied build_fake_container_fixture() from bib
This is a very sad commit, it copies the code to make a fake
container with a faked osbuild.

The bright side is that the plan is to rewrite these helpers
in go and then we can just have a shared testutil module that
can be imported from both bib/ibcli.

And/or the two repos will merge and its just a (simple) test
helper. But I do not feel good (at all) about this.