main: add new `--extra-repo-path` flag
This commit adds a  new flag `--extra-repo-path` that can be used
to point to a repository json file that is added to the base
repositories when depsolving.

This means that with a file like `extra.json`:
```json
{
  "x86_64": [
    {
      "name": "osbuild",
      "baseurl": "https://download.copr.fedorainfracloud.org/results/@osbuild/osbuild/centos-stream-$releasever-x86_64/",
      "check_gpg": false
    }
  ]
}
```
and a blueprint like:
```toml
[[packages]]
name = "osbuild"
```
a manifest is generated that gets osbuild from the copr repo:
```console
$ image-builder  --extra-repo-path ./extra.json  manifest qcow2 --distro centos-9 --blueprint ./bp.toml |jq|grep python3-osbuild
          "path": "08589574-osbuild/python3-osbuild-139-1.20250131112359008804.main.18.gd8f353b3.el9.noarch.rpm",
```
Note that this is part of the base repositories so anything with a
higher version number will get pulled from the extra-repo, even
system libraries or kernels.

Note as well that there is no safeguard right now against adding
extra repos for the wrong version of the distro, i.e. one could
add an extra repo build against/for fedora-42 on a fedora-40 image
which most likely will break with bad depsolve errors. But that
is okay, this option is meant for advanced users.