tools: provide a `run-in-mock` helper for the koji-plugin
This commit adds a `tools/run-in-mock` script that can be used
to wrap running image-builder inside a mock chroot. Without
this wrapper bwrap will error because it expects a "/" directory
but that is missing in the mock chroots.

Kudos to jlebon and his mock-mount.sh that is available here:
https://gist.github.com/jlebon/fb6e7c6dcc3ce17d3e2a86f5938ec033

See also https://github.com/containers/bubblewrap/issues/135