entrypoint.sh: detect rootful/privileged containers
This commit detects if a container is rootful and privileged and
only sets up the selinux/osbuild workarounds when that is the case.
With that change we can run the container as an unprivileged user
and use "supermin" inside to do the building while still supporting
the existing workflow (i.e. direct building inside the container
if we are privileged and rootful).