progress: error if there are not enough priviledges to run osbuild
This commit detects if there are enough priviledges to run osbuild
and if not errors. This prepares for the `supermin` build mode
which will be used in the followup commits when not enough
priviledges are detected.

It currently checks if it can do a mknod for the /dev/sda1 block
devices.

Do a functional check here as this is the most reliable way
to see if we have enough privileges. Checking for CAP_SYS_ADMIN
is not good enough as that is available in a privileged user
container but has not enough privs to create device nodes or
mount filesystems.

Being able to check for mknod is sufficient, alternatively
we could try to mount ext4 but that is much more involved
as we would need to do the loop device dance and call
mkfs.* which may not be available (or we would just try
syscall.Mount("/any/block-device", "/valid/mnt", 0, "")
and return false if we get EPERM  there.

Thanks for Ondrej for pointing this out!