progress: add `IMAGE_BUILDER_EXPERIMENTAL=supermin` rootless builds
This adds support to build images without root by using `supermin`
and a virtual machine. With `supermin` we can create a minimal
VM with osbuild and all supported pacakge very quickly and then
map /store and /output in.

This works on a regular system with a uid != 0 and /dev/kvm access
and inside a container that has access to /dev/kvm.

There are a few wrinkles:
- /dev/kvm must be available for the VM or things are really slow
- /store cannot be mounted from the host and mounted directly because
  there will be selinux label/uid map issues between VM and host
- progress reporting is currently not supported (but adding it
  should be conceptually not very hard, see todo in the code).