supermin: add build done stamp file
This commit adds a stamp file if the build is successful. This is
needed because it is non-trivial to exit qemu non-zero [0] if osbuild
fails, so instead we just generate a build-done "stamp" file in
output and if that is missing we know something went wrong and show
the log. It has to be a random stamp to avoid having to deal
with pre-existing build-stamp files.

[0] It is non-trivial because any failure inside the VM will make
the kernel panic as pid1 dies. In this case we need to stop the
VM and shutdown so that the VM does not hang forever. But shutdown
via non-zero exit code needs very osbsure ISA debug ports that
we could poke from inside the VM. Given that we are only interested
in two cases:
1. build successful
2. build failed
and if the build failed we need to report a log file just using
the stamp seems the most expedient option.