bib: simplify extractTLSKeys()
The extractTLSKeys() takes a SimplefileReader so that its
testable. The downside is that it makes it slightly harder
to use the helper because of the extra parameter.

But we don't really need this, we can make the testing
entirely invisible for the API user via the common
```go
var osReadFile = os.ReadFile
```
pattern.

This commit is doing exactly this.