progress: allow printing after progress is done
When something needs to be printed after the progress bar is done, it
was not possible to do so without messing up the terminal. This patch
adds a mutex around the buffer used to store output while the progress
bar is active. This allows to safely print after stopping the progress
bar.