main: add `--bootc-build-ref` option to set build container
Some use-cases want to use a "build-container" for the disk
image generation. One example is e.g. very restricted images
that do not even ship with python. Here in order to build
the image one needs to be able to specify a different container
that can act as the build-root. Its up to the user to ensure
that the build-container has a sufficiently compatible userland
for tools like e.g. mkfs for the target containers kernel.