legacy_iso: update lorax handling for images PR#1886
With the merge of https://github.com/osbuild/images/pull/1886
the handling of the lorax templates changed. Ideally we would
follow images here closer and use the distro YAML loader to
load the metadata from images:data/distrodefs. But as a quick
fix this just duplicates the logic we had before.