main: embed manifestgen.options into manifestOptions
When we generate the osbuild manifest we need to take a bunch
of commandline options into account. These are collected and
passed via `manifestOptions` in the CLI. There is a (small)
overlap with options that are then "converted" to options that
need to be passed to `manifestgen` via `manifestgen.Options`.

Previously there was manual code for this but its slightly
nicer to just embedd the `manifestgen.Options` into the
more general `manifestOptions` of the CLI. This avoid some
boilerplate code and might serve as a useful pattern in other
places. So this commit does that now (even though the wins
are not that big).