vm: drop paramiko
And replace with ssh/scp binaries. This is probably more
robust than paramiko and more "standard". As a side effect
we do no longer request an explicit tty. We could force
that with "-t" in ssh but we did that with paramiko and
that caused issues that e.g. "systemctl" would detect
a tty and go into pager mode. So lets start without and
we can always add "-t" to our ssh invocation to force
it.