repos: don't include builtins
With the rename of `data-dir` to `force-data-dir` it is implied that
only data within the `force-data-dir` is being used. However, the
repositories were always merged with the built-in embedded filesystem.

This commit makes it so that when a `force-data-dir` is passed as an
argument then only the repositories inside that directory are used.

Note that the repository files inside the data directory still must
match any defined distributions in the definitions.

Also add a small smoke test that uses `force-data-dir` to only set up a
`rhel-10.0` repository and assert that there are only listed images for
`rhel-10.0`.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>