github: add manifest test in osbuild container
In oder to test if we can get subscribed content inside
a ibcli container we need a subscribed host. This is tricky
with github because:
a) there are no `runs-on: {fedora,centos,rhel}` systems
b) creating a fedora/centos chroot and running podman in it
   is tricky because podman really dislikes chroots

So this commit runs it inside the osbuild test container that
is fedora based and relies on podman-in-podman container
execution (for this we need to run with `--privileged`).