main: add a boostrap container if target-distro != host-distro
This commit should mitigate issue with incompatible rpm
versions. It is a bit annoying that this is now generating
host dependant manifests so the final version might be
to just drop the check and *always* insert the bootstrap
pipeline as we do not know (in general) on what target
system the build will happen. Even when called with
`build` the user could still also use `--with-manifest`
and then later build the same manifest on a different
machine. Its also nice in that we remove one more host
variable seeping into the build.

This fixes our side of
https://issues.redhat.com/browse/RHEL-128741