Containerfile.bib: adjust for the ibcli environment
Adjust the bib containerfile to the environment it lives in
in ibcli. This is not ideal as we now have drift between
ibcli/bib version of the Containerfile. As much of this
should be merged back into bib so that ideally we would just
copy the files and even have a test that errors if there is
drift.