bib: document details about {,bib}cmd{ManifestFromCobra,Build}
The {,bib}CmdBuild and {,bib}ManifestFromCobra functions are
quite similar which raises the question what could be done
to share more code.

This commit documents the differences and why its hard to
consolidate them. Note that if we would add support for
anaconda-iso into bib via something like:
https://github.com/osbuild/image-builder-cli/compare/main...mvo5:consolidate-bib-anaconda-iso-wip?expand=1
we could remove some of this duplication, however this would
make the "generic" code a bit less nice (trade-offs!).

But other differences like the lack of renaming of the produced
results or the multi-export builds in bib are harder to
consolidate and probably not worth it (i.e. the generic code
becomes more ugly than the de-duplication niceness wins).