test: move bib tests from bootc-image-builder
Move the integration tests from bootc-image-builder into
image-builder. With bootc-image-builder being merged into
image-builder we still want to run the integration tests
for bib because our goal is 100% compatibility. Note that
this does not preserve history. There is a PR [0] that
does that and while I'm a big fan in general I am not
sure its worth the trouble because GH will not allow us
to merge the repos and we need to manually force push
it. The history of the tests is still fully available
in the bootc-image-builder repo in cases it is needed.

[0] https://github.com/osbuild/image-builder-cli/pull/398