Containerfile.bib: move from bootc-image-builder
Similar as the previous commit, this moves the bootc-image-builder
Containerfile into the image-builder repository. We need it
so that we can run the tests and to (long term) move the
building of the bib konflux/GH containers into image-builder.

For now it will only be used in the integration tests and the
container builds for bib still happen in
github.com/osbuild/bootc-image-builder.

We will also need to add a test that ensures that the two
files keep in sync (igorning the small differences that will
be needed).