workflow: fix lint errors in bibtests.yaml
We run a stricter linting in image-builder-cli than we do
in bootc-image-builder. So the bibtests.yaml file needs
some adjustments. Usually I would just fold this into one
of the previous commit but as it moves the indent level it is
confusing if merged into the other one so this is its own.